feat: convert existing monolith to a multi-module monolith
