+++ Dockerfile
@@ -1,4 +1,4 @@
 FROM eclipse-temurin:17-jdk
-ARG JAR_FILE=build/libs/*.jar
+ARG JAR_FILE=module-api/build/libs/*.jar
 COPY ${JAR_FILE} app.jar
-ENTRYPOINT ["java","-jar","/app.jar"]
+ENTRYPOINT ["java", "-jar", "/app.jar"]
