+++ module-api/build.gradle
@@ -1,0 +1,13 @@
+
+bootJar.enabled = true
+jar.enabled = false
+
+dependencies {
+    implementation project(':module-core')
+    implementation project(':module-domain')
+
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+}
+
+tasks.register("prepareKotlinBuildScriptModel") {}
