+++ module-api/src/main/java/app/application/RegexProfanityFilter.java
