+++ module-api/src/main/java/app/ui/ProfanityController.java
