+++ module-core/src/main/java/app/core/exception/ErrorResponse.java
@@ -1,0 +1,14 @@
+package app.core.exception;
+
+import lombok.Builder;
+import org.springframework.http.HttpStatus;
+
+import java.time.LocalDateTime;
+
+@Builder
+public record ErrorResponse(
+        String message,
+        HttpStatus status,
+        LocalDateTime timestamp
+) {
+}
