+++ module-core/src/main/resources/application.yml
@@ -1,0 +1,14 @@
+spring:
+  profiles:
+    default: local
+---
+spring:
+  config:
+    activate:
+      on-profile: local,prod
+
+server:
+  servlet:
+    encoding:
+      charset: UTF-8 # 서버 인코딩 설정
+      force: true # 서버 인코딩 강제 설정
