+++ module-domain/build.gradle
@@ -1,0 +1,13 @@
+dependencies {
+    implementation project(':module-core')
+
+    //database
+    api 'org.springframework.boot:spring-boot-starter-data-jpa'
+    runtimeOnly 'com.mysql:mysql-connector-j'
+    runtimeOnly 'com.h2database:h2'
+
+    // ahocorasick
+    implementation 'org.ahocorasick:ahocorasick:0.6.3'
+}
+
+tasks.register("prepareKotlinBuildScriptModel") {}
