+++ module-domain/src/main/java/app/infra/JpaProfanityRepository.java
