+++ module-domain/src/test/java/app/moduledomain/ModuleDomainApplicationTests.java
@@ -1,0 +1,8 @@
+package app.moduledomain;
+
+class ModuleDomainApplicationTests {
+
+    void contextLoads() {
+    }
+
+}
