--- src/main/java/app/ProfanityFilterApplication.java
@@ -1,36 +1,0 @@
-package app;
-
-import com.zaxxer.hikari.HikariDataSource;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Primary;
-
-import javax.sql.DataSource;
-
-@SpringBootApplication
-public class ProfanityFilterApplication {
-
-    @Bean(name = "dataSource")
-    @Primary
-    @ConfigurationProperties("spring.datasource.hikari")
-    public DataSource dataSource() {
-        return DataSourceBuilder.create()
-                .type(HikariDataSource.class)
-                .build();
-    }
-
-    @Bean(name = "secondDataSource")
-    @ConfigurationProperties("second.datasource.hikari")
-    public DataSource secondDataSource() {
-        return DataSourceBuilder.create()
-                .type(HikariDataSource.class)
-                .build();
-    }
-
-    public static void main(String[] args) {
-        SpringApplication.run(ProfanityFilterApplication.class, args);
-    }
-}
