--- src/main/java/app/domain/isUsed.java
@@ -1,16 +1,0 @@
-package app.domain;
-
-public enum isUsed {
-    Y("used"),
-    N("not used");
-
-    private final String description;
-
-    isUsed(String description) {
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-}
